requirements.txt
# Core Dependencies
flask>=2.3.0,<3.0.0
werkzeug>=2.3.0,<3.0.0
gunicorn>=21.0.0

# Machine Learning
tensorflow>=2.13.0,<3.0.0
numpy>=1.24.0,<2.0.0
pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0